+++ waffle-chai/src/matchers/changeBalance.ts
@@ -1,6 +1,7 @@
 import {BigNumber, BigNumberish} from 'ethers';
 import {Account, getAddressOf} from './misc/account';
 import {getBalanceChange} from './changeEtherBalance';
+import {callPromise} from '../call-promise';
 
 export function supportChangeBalance(Assertion: Chai.AssertionStatic) {
   Assertion.addMethod('changeBalance', function (
@@ -8,25 +9,32 @@ export function supportChangeBalance(Assertion: Chai.AssertionStatic) {
     account: Account,
     balanceChange: BigNumberish
   ) {
-    const subject = this._obj;
-    const derivedPromise = Promise.all([
-      getBalanceChange(subject, account, {includeFee: true}),
-      getAddressOf(account)
-    ]).then(
-      ([actualChange, address]) => {
-        this.assert(
-          actualChange.eq(BigNumber.from(balanceChange)),
-          `Expected "${address}" to change balance by ${balanceChange} wei, ` +
-          `but it has changed by ${actualChange} wei`,
-          `Expected "${address}" to not change balance by ${balanceChange} wei,`,
-          balanceChange,
-          actualChange
-        );
+    callPromise(this);
+    const isNegated = this.__flags.negate === true;
+    const derivedPromise = this.callPromise.then(() => {
+      if (!('txResponse' in this)) {
+        throw new Error('The changeBalance matcher must be called on a transaction');
       }
-    );
+      return Promise.all([
+        getBalanceChange(this.txResponse, account, {includeFee: true}),
+        getAddressOf(account)
+      ]);
+    }).then(([actualChange, address]: [BigNumber, string]) => {
+      const isCurrentlyNegated = this.__flags.negate === true;
+      this.__flags.negate = isNegated;
+      this.assert(
+        actualChange.eq(BigNumber.from(balanceChange)),
+        `Expected "${address}" to change balance by ${balanceChange} wei, ` +
+        `but it has changed by ${actualChange} wei`,
+        `Expected "${address}" to not change balance by ${balanceChange} wei,`,
+        balanceChange,
+        actualChange
+      );
+      this.__flags.negate = isCurrentlyNegated;
+    });
     this.then = derivedPromise.then.bind(derivedPromise);
     this.catch = derivedPromise.catch.bind(derivedPromise);
-    this.promise = derivedPromise;
+    this.callPromise = derivedPromise;
     return this;
   });
 }