+++ waffle-chai/src/matchers/changeTokenBalance.ts
@@ -1,4 +1,5 @@
-import {BigNumber, BigNumberish, Contract} from 'ethers';
+import {BigNumber, BigNumberish, Contract, providers} from 'ethers';
+import {callPromise} from '../call-promise';
 import {Account, getAddressOf} from './misc/account';
 
 export function supportChangeTokenBalance(Assertion: Chai.AssertionStatic) {
@@ -8,37 +9,50 @@ export function supportChangeTokenBalance(Assertion: Chai.AssertionStatic) {
     account: Account,
     balanceChange: BigNumberish
   ) {
-    const subject = this._obj;
-    const derivedPromise = Promise.all([
-      getBalanceChange(subject, token, account),
-      getAddressOf(account)
-    ]).then(
-      ([actualChange, address]) => {
-        this.assert(
-          actualChange.eq(BigNumber.from(balanceChange)),
-          `Expected "${address}" to change balance by ${balanceChange} wei, ` +
-          `but it has changed by ${actualChange} wei`,
-          `Expected "${address}" to not change balance by ${balanceChange} wei,`,
-          balanceChange,
-          actualChange
-        );
+    callPromise(this);
+    const isNegated = this.__flags.negate === true;
+    const derivedPromise = this.callPromise.then(async () => {
+      if (!('txReceipt' in this)) {
+        throw new Error('The changeTokenBalance matcher must be called on a transaction');
       }
-    );
+      const address = await getAddressOf(account);
+      const actualChanges = await getBalanceChange(this.txReceipt, token, address);
+      return [actualChanges, address];
+    }).then(([actualChange, address]: [BigNumber, string]) => {
+      const isCurrentlyNegated = this.__flags.negate === true;
+      this.__flags.negate = isNegated;
+      this.assert(
+        actualChange.eq(BigNumber.from(balanceChange)),
+        `Expected "${address}" to change balance by ${balanceChange} wei, ` +
+          `but it has changed by ${actualChange} wei`,
+        `Expected "${address}" to not change balance by ${balanceChange} wei,`,
+        balanceChange,
+        actualChange
+      );
+      this.__flags.negate = isCurrentlyNegated;
+    });
     this.then = derivedPromise.then.bind(derivedPromise);
     this.catch = derivedPromise.catch.bind(derivedPromise);
-    this.promise = derivedPromise;
+    this.callPromise = derivedPromise;
     return this;
   });
 }
 
 async function getBalanceChange(
-  transactionCall: (() => Promise<void> | void),
+  txReceipt: providers.TransactionReceipt,
   token: Contract,
-  account: Account
+  address: string
 ) {
-  const balanceBefore: BigNumber = await token['balanceOf(address)'](await getAddressOf(account));
-  await transactionCall();
-  const balanceAfter: BigNumber = await token['balanceOf(address)'](await getAddressOf(account));
+  const txBlockNumber = txReceipt.blockNumber;
+
+  const balanceBefore: BigNumber = await token['balanceOf(address)'](
+    address,
+    {blockTag: txBlockNumber - 1}
+  );
+  const balanceAfter: BigNumber = await token['balanceOf(address)'](
+    address,
+    {blockTag: txBlockNumber}
+  );
 
   return balanceAfter.sub(balanceBefore);
 }