+++ waffle-chai/src/matchers/emit.ts
@@ -1,4 +1,5 @@
 import {Contract, providers, utils} from 'ethers';
+import {callPromise} from '../call-promise';
 import {waitForPendingTransaction} from './misc/transaction';
 
 export function supportEmit(Assertion: Chai.AssertionStatic) {
@@ -7,15 +8,22 @@ export function supportEmit(Assertion: Chai.AssertionStatic) {
       .filter((log) => log.address && log.address.toLowerCase() === contractAddress.toLowerCase());
 
   Assertion.addMethod('emit', function (this: any, contract: Contract, eventName: string) {
-    const tx = this._obj;
-    const isNegated = this.__flags.negate === true;
-    if (!('promise' in this)) {
-      this.promise = waitForPendingTransaction(tx, contract.provider)
-        .then((receipt) => { this.receipt = receipt; });
+    if (typeof this._obj === 'string') {
+      // Handle specific case of using transaction hash to specify transaction. Done for backwards compatibility.
+      this.callPromise = waitForPendingTransaction(this._obj, contract.provider)
+        .then(txReceipt => {
+          this.txReceipt = txReceipt;
+        });
+    } else {
+      callPromise(this);
     }
-    this.promise = this.promise
+    const isNegated = this.__flags.negate === true;
+    this.callPromise = this.callPromise
       .then(() => {
-        const receipt: providers.TransactionReceipt = this.receipt;
+        if (!('txReceipt' in this)) {
+          throw new Error('The emit matcher must be called on a transaction');
+        }
+        const receipt: providers.TransactionReceipt = this.txReceipt;
         let eventFragment: utils.EventFragment | undefined;
         try {
           eventFragment = contract.interface.getEvent(eventName);
@@ -39,78 +47,24 @@ export function supportEmit(Assertion: Chai.AssertionStatic) {
         }
 
         const topic = contract.interface.getEventTopic(eventFragment);
-        this.logs = filterLogsWithTopics(receipt.logs, topic, contract.address);
+        this.args = filterLogsWithTopics(receipt.logs, topic, contract.address);
         // As this callback will be resolved after the chain of matchers is finished, we need to
         // know if the matcher has been negated or not. To simulate chai behaviour, we keep track of whether
         // the matcher has been negated or not and set the internal chai flag __flags.negate to the same value.
         // After the assertion is finished, we set the flag back to original value to not affect other assertions.
         const isCurrentlyNegated = this.__flags.negate === true;
         this.__flags.negate = isNegated;
-        this.assert(this.logs.length > 0,
+        this.assert(this.args.length > 0,
           `Expected event "${eventName}" to be emitted, but it wasn't`,
           `Expected event "${eventName}" NOT to be emitted, but it was`
         );
         this.__flags.negate = isCurrentlyNegated;
       });
-    this.then = this.promise.then.bind(this.promise);
-    this.catch = this.promise.catch.bind(this.promise);
+    this.then = this.callPromise.then.bind(this.callPromise);
+    this.catch = this.callPromise.catch.bind(this.callPromise);
     this.contract = contract;
     this.eventName = eventName;
-    return this;
-  });
-
-  const assertArgsArraysEqual = (context: any, expectedArgs: any[], log: any) => {
-    const actualArgs = (context.contract.interface as utils.Interface).parseLog(log).args;
-    context.assert(
-      actualArgs.length === expectedArgs.length,
-      `Expected "${context.eventName}" event to have ${expectedArgs.length} argument(s), ` +
-      `but it has ${actualArgs.length}`,
-      'Do not combine .not. with .withArgs()',
-      expectedArgs.length,
-      actualArgs.length
-    );
-    for (let index = 0; index < expectedArgs.length; index++) {
-      if (expectedArgs[index]?.length !== undefined && typeof expectedArgs[index] !== 'string') {
-        for (let j = 0; j < expectedArgs[index].length; j++) {
-          new Assertion(actualArgs[index][j]).equal(expectedArgs[index][j]);
-        }
-      } else {
-        if (actualArgs[index].hash !== undefined && actualArgs[index]._isIndexed === true) {
-          const expectedArgBytes = utils.isHexString(expectedArgs[index])
-            ? utils.arrayify(expectedArgs[index]) : utils.toUtf8Bytes(expectedArgs[index]);
-          new Assertion(actualArgs[index].hash).to.be.oneOf(
-            [expectedArgs[index], utils.keccak256(expectedArgBytes)]
-          );
-        } else {
-          new Assertion(actualArgs[index]).equal(expectedArgs[index]);
-        }
-      }
-    }
-  };
-
-  const tryAssertArgsArraysEqual = (context: any, expectedArgs: any[], logs: any[]) => {
-    if (logs.length === 1) return assertArgsArraysEqual(context, expectedArgs, logs[0]);
-    for (const index in logs) {
-      try {
-        assertArgsArraysEqual(context, expectedArgs, logs[index]);
-        return;
-      } catch {}
-    }
-    context.assert(false,
-      `Specified args not emitted in any of ${context.logs.length} emitted "${context.eventName}" events`,
-      'Do not combine .not. with .withArgs()'
-    );
-  };
-
-  Assertion.addMethod('withArgs', function (this: any, ...expectedArgs: any[]) {
-    if (!('promise' in this)) {
-      throw new Error('withArgs() must be used after emit()');
-    }
-    this.promise = this.promise.then(() => {
-      tryAssertArgsArraysEqual(this, expectedArgs, this.logs);
-    });
-    this.then = this.promise.then.bind(this.promise);
-    this.catch = this.promise.catch.bind(this.promise);
+    this.txMatcher = 'emit';
     return this;
   });
 }