+++ waffle-chai/test/matchers/bigNumber.test.ts
@@ -141,32 +141,36 @@ describe('UNIT: BigNumber matchers', () => {
 
   describe('closeTo', () => {
     it('.to.be.closeTo', () => {
-      expect(BigNumber.from(100)).to.be.closeTo(BigNumber.from(101), 10);
-      expect(BigNumber.from(100)).to.be.closeTo(BigNumber.from(101), BigNumber.from(10));
+      checkAll(100, 101, (a, b) => expect(a).to.be.closeTo(b, 10));
+      checkAll(100, 101, (a, b) => expect(a).to.be.closeTo(b, BigNumber.from(10)));
     });
 
     it('.not.to.be.closeTo', () => {
-      expect(BigNumber.from(100)).not.to.be.closeTo(BigNumber.from(111), 10);
-      expect(BigNumber.from(100)).not.to.be.closeTo(BigNumber.from(111), BigNumber.from(10));
+      checkAll(100, 111, (a, b) => expect(a).not.to.be.closeTo(b, 10));
+      checkAll(100, 111, (a, b) => expect(a).not.to.be.closeTo(b, BigNumber.from(10)));
     });
 
     it('expect to throw on error', () => {
-      expect(() => expect(BigNumber.from(100)).to.be.closeTo(BigNumber.from(111), 10)).to.throw(
-        AssertionError,
-        'Expected "100" to be within 10 of 111'
-      );
-      expect(() => expect(BigNumber.from(100)).to.be.closeTo(BigNumber.from(111), BigNumber.from(10))).to.throw(
-        AssertionError,
-        'Expected "100" to be within 10 of 111'
-      );
-      expect(() => expect(BigNumber.from(100)).not.to.be.closeTo(BigNumber.from(101), 10)).to.throw(
-        AssertionError,
-        'Expected "100" NOT to be within 10 of 101'
-      );
-      expect(() => expect(BigNumber.from(100)).not.to.be.closeTo(BigNumber.from(101), BigNumber.from(10))).to.throw(
-        AssertionError,
-        'Expected "100" NOT to be within 10 of 101'
-      );
+      checkAll(100, 111, (a, b) => {
+        expect(() => expect(BigNumber.from(a)).to.be.closeTo(BigNumber.from(b), 10)).to.throw(
+          AssertionError,
+          'Expected "100" to be within 10 of 111'
+        );
+        expect(() => expect(BigNumber.from(a)).to.be.closeTo(BigNumber.from(b), BigNumber.from(10))).to.throw(
+          AssertionError,
+          'Expected "100" to be within 10 of 111'
+        );
+      });
+      checkAll(100, 101, (a, b) => {
+        expect(() => expect(BigNumber.from(a)).not.to.be.closeTo(BigNumber.from(b), 10)).to.throw(
+          AssertionError,
+          'Expected "100" NOT to be within 10 of 101'
+        );
+        expect(() => expect(BigNumber.from(a)).not.to.be.closeTo(BigNumber.from(b), BigNumber.from(10))).to.throw(
+          AssertionError,
+          'Expected "100" NOT to be within 10 of 101'
+        );
+      });
     });
   });
 });