+++ waffle-chai/test/matchers/chaining.test.ts
@@ -1,0 +1,6 @@
+import {chainingMatchersTest} from './chainingTest';
+import {describeMockProviderCases} from './MockProviderCases';
+
+describeMockProviderCases('INTEGRATION: chaining matchers', (provider) => {
+  chainingMatchersTest(provider);
+});