+++ waffle-hardhat/hardhat.config.ts
@@ -6,5 +6,17 @@ module.exports = {
       initialDate: '2020-01-01T00:00:00',
       allowUnlimitedContractSize: true,
     },
-  }
+  },
+  paths: {
+    sources: './contracts',
+    artifacts: './build',
+    cache: './cache',
+  },
+  solidity: {
+    compilers: [
+      {
+        version: "0.8.10"
+      }
+    ]
+  } 
 }