+++ waffle-chai/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @ethereum-waffle/chai
 
+## 4.0.0-alpha.24
+
+### Patch Changes
+
+- ba71ce4: Add support for custom errors in hardhat
+- 587ff49: Allow chaining matchers
+- 4622881: Support panic codes in ganache
+- Updated dependencies [4622881]
+  - @ethereum-waffle/provider@4.0.0-alpha.23
+
+## 4.0.0-alpha.23
+
+### Patch Changes
+
+- Updated dependencies [671f139]
+  - @ethereum-waffle/provider@4.0.0-alpha.22
+
+## 4.0.0-alpha.22
+
+### Patch Changes
+
+- e8adce7: 🧼 Add src folders to npm
+- Updated dependencies [e8adce7]
+  - @ethereum-waffle/provider@4.0.0-alpha.21
+
+## 4.0.0-alpha.21
+
+### Patch Changes
+
+- aa18bfa: Improve hardhat error catching
+
 ## 4.0.0-alpha.20
 
 ### Patch Changes